move legacy controller
