+++ webapp/static/to_compile/js/carte.ts
@@ -8,7 +8,7 @@ import ScrollController from "../controllers/shared/scroll_controller"
 import AddressAutocompleteController from "../controllers/carte/address_autocomplete_controller"
 import MapController from "../controllers/carte/map_controller"
 import SearchSolutionFormController from "../controllers/carte/search_solution_form_controller"
-import SsCatObjectAutocompleteController from "../controllers/carte/ss_cat_object_autocomplete_controller"
+import SsCatObjectAutocompleteController from "../controllers/formulaire/ss_cat_object_autocomplete_controller"
 import ActeurDetailsController from "../controllers/carte/acteur_details"
 import NextAutocompleteController from "../controllers/shared/next_autocomplete_controller"
 import ResponsiveController from "../controllers/carte/responsive_controller"
