+++ webapp/templates/ui/pages/carte.html
@@ -22,9 +22,7 @@
         data-action="
         map:updateBbox->search-solution-form#updateBboxInput
         search-solution-form:captureInteraction->analytics#captureInteractionWithMap
-        address-autocomplete:formSubmit->search-solution-form#submitForm
-        ss-cat-object-autocomplete:formSubmit->search-solution-form#submitForm
-        ss-cat-object-autocomplete:optionSelected->search-solution-form#checkSsCatObjetErrorForm"
+        address-autocomplete:formSubmit->search-solution-form#submitForm"
         class="
         qf-rounded-lg qf-border qf-border-solid qf-border-grey-900
         qf-group
